+++ src/main/java/com/acco/life/entity/AccountTransaction.java
@@ -24,13 +24,13 @@ public class AccountTransaction extends BaseColumnEntity {
      * 所属用户
      */
     @Column("user_id")
-    private Integer userId;
+    private Long userId;
 
     /**
      * 发生账户
      */
     @Column("account_id")
-    private Integer accountId;
+    private Long accountId;
 
     /**
      * 交易类型：1-income，2-expense，3-transfer_out，4-transfer_in
@@ -48,13 +48,13 @@ public class AccountTransaction extends BaseColumnEntity {
      * 分类ID
      */
     @Column("category_id")
-    private Integer categoryId;
+    private Long categoryId;
 
     /**
      * 关联交易ID
      */
     @Column("related_transaction_id")
-    private Integer relatedTransactionId;
+    private Long relatedTransactionId;
 
     /**
      * 摘要说明
@@ -84,5 +84,5 @@ public class AccountTransaction extends BaseColumnEntity {
      * 账单归属ID
      */
     @Column("statement_id")
-    private Integer statementId;
+    private Long statementId;
 }
